feat: implement carousel section detection
this adds a way of detecting which section (team, contact, ..) is
currently being viewed, for this the Intersection Observer API is being
used via the `react-intersection-observer` package.
using this we determine the active section and forward this info to the
carousel component.
what is still missing is using "scrollIntoView(..)" when clicking on an
item in the carousel component.
as all sections have html ids attached to them we have somewhat similar
behaviour already as the carousel currently contains links which have
`#<id>` as their href prop.